Creating a starter component and fixing some unified errors.
The component content needs to be reviewed.
 - Proper icons are needed
 - The screenshots should be 740 x 400
 - The Details.md is cut from the readme
 - The GettingStarted.md is also cut from the readme

As for the unified project, more work is needed

I simply got the project to compile.  There are still differences in the
ApiDefinitioins files.  Methods are missing in the unified version.
Also an audit should be done to verify that int are to be used in the unified version
and not nint.